Split color into separate lib
